refactor de movies y estilos aaplicados
